copy/paste of the legacy switch
I put in my if statement to detect if legacy mode is on in several different command files, and consolidated the legacy command into the standard default commandsets file.

Working on getting +sheet to work with legacy mode, which is proving to be a challenge. Updates to come for that, but so far so good. This has now incorporated the basics for Vampire the Requiem 1st edition.